Move split mesh metadata checks out of MooseApp
Introduce a SplitMeshMetaData helper to own split mesh input summaries, fingerprints, metadata writes, and validation. Keep MooseApp limited to generic restartable metadata services and update split mesh write/read call sites to use the helper directly.

Use the existing MeshMetaData/split_mesh namespace for the stored fingerprint and summary, add validation-flow comments, make the diagnostic summary resemble mesh input syntax, and narrow the metadata-edit test helper to the no-fingerprint compatibility case.
