+++ framework/src/mesh/FileMesh.C
@@ -13,6 +13,7 @@
 #include "Moose.h"
 #include "MooseApp.h"
 #include "FileMeshGenerator.h"
+#include "SplitMeshMetaData.h"
 
 #include "libmesh/exodusII_io.h"
 #include "libmesh/mesh_tools.h"
@@ -147,7 +148,7 @@ FileMesh::buildMesh()
         MeshTools::clear_spline_nodes(getMesh());
 
       if (_is_split)
-        _app.checkSplitMeshMetaData(_file_name);
+        SplitMeshMetaData(_app).check(_file_name);
 
       _app.possiblyLoadRestartableMetaData(MooseApp::MESH_META_DATA, _file_name);
